rename types files, change state type (add blocks: key)
